Float é o certo aqui.

+++ Mundo 01 - Fundamentos/Exercicios/ex007.py
@@ -1,15 +1,15 @@
 #-------------------------------------------------------------
-n1 = int(input('Nota 1: '))
-n2 = int(input('Nota 2: '))
-n3 = int(input('Nota 3: '))
-n4 = int(input('Nota 4: '))
+n1 = float(input('Nota 1: '))
+n2 = float(input('Nota 2: '))
+n3 = float(input('Nota 3: '))
+n4 = float(input('Nota 4: '))
 m = (n1+n2+n3+n4) / 4
 print(f'A média = {m:.2f}') # formatado com duas casas decimais;
 #-------------------------------------------------------------
 print('\n')
 #-------------------------------------------------------------
-v1 = int(input('Prova 1: '))
-v2 = int(input('Prova 2: '))
+v1 = float(input('Prova 1: '))
+v2 = float(input('Prova 2: '))
 print(f'A média = {(v1+v2)/2:.2f}')
